Commons: - Added to TaskManager the capability to interrupt running Tasks execution on configuration reloading.   Modified TaskManager in order to append to Scheduler's Threads name also the executing TaskGroup.TaskName in order to simplify troubleshooting. - Corrected JSON<->XML numeric values conversion. - Removed XPath engine re-initialization during startup. - Removed recursive invocation of JMXEntrypoint.init() during Initializers execution. - Improved (N)CLOB fields management in sqlX metadata. - Improved reload configuration locking in order to avoid dead-lock. - Improved JMX entry-point initialization.
Core:
- Corrected @loggerLevel handling in GVCoreCallNode, GVIteratorOperationNode/CoreCall and parallel nodes
- Added to GVCoreCallNode and GVIteratorOperationNode/CoreCall a new attibute change-log-master-service.
  If true change the log master service with the used one (configured or calculated), in order to write log messages into a different log file.
  Used only if @change-log-context is true.
- Improved init/run logging of ServiceCallerTask.NextAction status
- Corrected GVCore pool element clean-up after configuration reload.
- Moved to INFO level the end trace message of workflow nodes, added execution time in ms.
- Enabled missing FTPS remote manager in dtds
- Improved fail-fast capability on XA abort in transacted workflows execution
- Improved transaction error management in workflow engine
- Corrected GVSubFlowCallNode @dynamic behavior
- Modified Core's configuration reload management, in order to better handle already running service instances.
- Added ACL check to GVCoreCallNode and IteratorController in order to enforce ACL check across flow calls.
- Added Input/Output services support to GVSubFlow* and GVCoreCallNode nodes.
- Added JMX interface to subflows (statistics and executing instances status)
- Added support to interruptible services
    - Added JMX operations that enables the interruption of Operations and/or SubFlows (by ID and Thread name)

DataHandler:
- Provided a new getData method on DH retrievers in order to specify the parameter's actual values separator;
  useful if the default separator (comma) is present into the parameter's actual values
- Improved handling of NCHAR,NVARCHAR,NCLOB fields in DataHandler.
- Corrected NullPointerException in DH DBOInsert/DBOUpdate without an input XML.

DTE:
- Changed cardinality of ToPreload DTE element.
- Improved empty array management in XML to JSON conversion.
- Improved management of multi-root JSON document in JSON <-> XML simple conversion policy.
- Improved namespaces exclusion in JSON <-> XML simple conversion policy and added a BadgerFish conversion policy.
- Moved to a GV specific version of org.json json library: now the properties creation order is keep in dump and properties listing, in order of better handle conversion to/from XML documents.
- Improved DTEControlled clean-up on destroy.

Task Scheduler:
- Improved TaskManager reload configuration locking in order to avoid dead-lock.

Adapter/VCL Axis2:
- Corrected mixed SOAP 1.1/1.2 request/response handling in Axis2MessageContextDataProvider.
- Added new property WS_REQ_SOAP_VERSION to input GVBuffer.
- Added optional attribute to ws-call to avoid to set SOAPAction header in the request

Adapter HTTP Inbound:
- Added management of output GVBuffer property HTTP_RESP_FILE_NAME to RESTHttpServletMapping in order to define the name of the file to be downloaded through an HTTP request.

VCL File:
- Added to RemoteManager methods the capability to provide at the real implementations a Map containing implementation-specific properties.
  Updated the remote commands to reads a list of properties from configuration.
- Corrected zip file closing on unzip operation.

JMS Listener:
- Improved listeners instantiation algorithm in JMS listener pool.
- Improved Dump of JMS's MapMessage.
- Improved debug logging in JMSForwardListener

VCL HTTP:
- Corrected RegisterHttpClientProtocolManager initialization, in order to avoid double JMX registration.
- Added new output property to http-call VCL plug-in: GVHTTP_RESPONSE_MESSAGE containing the response status text.
- Added to http-call the capability to set a context path before the method URI

VCL J2EE:
- Added to JMS Enqueue/Dequeue VCL operations the capability to exclude:
    - GVBuffer -> Message properties enrichment during enqueue
    - Message -> GVBuffer properties enrichment during dequeue
- Added a new attribute to jms-dequeue to specifies if the dequeue must keep input GVBuffer properties, not present in the JMS message, in the output GVBuffer.

+++ gvesb-core/gvcore/src/main/java/it/greenvulcano/gvesb/core/flow/GVSubFlow.java
@@ -23,16 +23,17 @@ import it.greenvulcano.configuration.XMLConfig;
 import it.greenvulcano.configuration.XMLConfigException;
 import it.greenvulcano.gvesb.buffer.GVBuffer;
 import it.greenvulcano.gvesb.core.config.InvocationContext;
+import it.greenvulcano.gvesb.core.config.ServiceConfigManager;
 import it.greenvulcano.gvesb.core.debug.DebugSynchObject;
 import it.greenvulcano.gvesb.core.debug.DebuggingInvocationHandler;
 import it.greenvulcano.gvesb.core.debug.ExecutionInfo;
 import it.greenvulcano.gvesb.core.exc.GVCoreConfException;
 import it.greenvulcano.gvesb.core.exc.GVCoreException;
-import it.greenvulcano.gvesb.core.jmx.OperationInfo;
 import it.greenvulcano.gvesb.core.jmx.ServiceOperationInfoManager;
+import it.greenvulcano.gvesb.core.jmx.SubFlowInfo;
 import it.greenvulcano.gvesb.gvdte.controller.DTEController;
-import it.greenvulcano.gvesb.internal.GVInternalException;
 import it.greenvulcano.gvesb.log.GVFormatLog;
+import it.greenvulcano.gvesb.statistics.StatisticsDataManager;
 import it.greenvulcano.log.GVLogger;
 import it.greenvulcano.util.xpath.XPathFinder;
 
@@ -68,9 +69,18 @@ public class GVSubFlow
      */
     private String                  flowName       = "";
     /**
-     * the jmx operation info instance
+     * the jmx subflow info instance
      */
-    private OperationInfo           operationInfo  = null;
+    private SubFlowInfo             subflowInfo    = null;
+
+    /**
+     * Used to get an GVServiceConf of a specific service (SYSTEM + SERVICE).
+     */
+    private ServiceConfigManager  gvSvcConfMgr          = null;
+    /**
+     * The Statistics StatisticsDataManager to be used.
+     */
+    private StatisticsDataManager statisticsDataManager = null;
 
     private DTEController           dteController  = null;
 
@@ -123,17 +133,7 @@ public class GVSubFlow
 
         this.isSingleThread = isSingleThread;
 
-        if (!isSingleThread) {
-            String dteConfFileName = "GVDataTransformation.xml";
-            logger.debug("DTE configuration file: " + dteConfFileName + ".");
-            try {
-                dteController = new DTEController(dteConfFileName);
-            }
-            catch (Exception exc) {
-                logger.error("Error initializing DTEController from file: " + dteConfFileName, exc);
-            }
-            gvContext = new InvocationContext();
-        }
+        initInvocationContext();
 
         initFlowNodes(gvsfNode);
 
@@ -151,6 +151,12 @@ public class GVSubFlow
         Level level = null;
         try {
             level = GVLogger.setThreadMasterLevel(loggerLevel);
+            try {
+                setInvocationContext(null);
+            }
+            catch (Exception exc) {
+                logger.error("Error initializing context", exc);
+            }
 
             try {
                 firstNode = XMLConfig.get(instNode, "@first-node");
@@ -187,6 +193,7 @@ public class GVSubFlow
             }
         }
         finally {
+            resetInvocationContext();
             GVLogger.removeThreadMasterLevel(level);
         }
     }
@@ -199,8 +206,10 @@ public class GVSubFlow
      * @return the output data
      * @throws GVCoreException
      *         if errors occurs
+     * @throws InterruptedException
+     *         if the current Thread is interrupted
      */
-    public GVBuffer perform(GVBuffer gvBuffer, boolean onDebug) throws GVCoreException
+    public GVBuffer perform(GVBuffer gvBuffer, boolean onDebug) throws GVCoreException, InterruptedException
     {
         onDebug = onDebug || "true".equalsIgnoreCase(gvBuffer.getProperty("GV_FLOW_DEBUG"));
         if (!onDebug) {
@@ -209,32 +218,80 @@ public class GVSubFlow
         }
 
         try {
-            if (!isSingleThread) {
-                InvocationContext mainCtx = (InvocationContext) InvocationContext.getInstance();
-                gvContext.setContext(flowName, gvBuffer);
-                gvContext.setGVServiceConfigManager(mainCtx.getGVServiceConfigManager());
-                gvContext.setStatisticsDataManager(mainCtx.getStatisticsDataManager());
-                // gvContext.setExtraField("DTE_CONTROLLER",
-                // mainCtx.getExtraField("DTE_CONTROLLER"));
-                gvContext.setExtraField("DTE_CONTROLLER", dteController);
-            }
-        }
-        catch (GVInternalException exc) {
-            throw new GVCoreException("GVCORE_FLOW_EXCEPTION_ERROR", new String[][]{{"operation", flowName}}, exc);
-        }
+            setInvocationContext(gvBuffer);
 
-        try {
-            if (!isSingleThread) {
-                gvContext.push();
-            }
             GVBuffer outData = internalPerform(gvBuffer, onDebug);
             return outData;
         }
         catch (GVCoreException exc) {
             throw exc;
         }
+        catch (InterruptedException exc) {
+            throw exc;
+        }
         finally {
             cleanUp();
+        }
+    }
+
+    public String getFlowName() {
+        return this.flowName;
+    }
+
+    /**
+     * @param gvBuffer
+     * @throws GVCoreException
+     */
+    private void initInvocationContext() throws GVCoreConfException {
+        if (isSingleThread) {
+            return;
+        }
+        statisticsDataManager = new StatisticsDataManager();
+        try {
+            statisticsDataManager.init();
+        }
+        catch (Exception exc) {
+            logger.error("Error initializing Statistics Manager", exc);
+        }
+        String dteConfFileName = "GVDataTransformation.xml";
+        logger.debug("DTE configuration file: " + dteConfFileName + ".");
+        try {
+            dteController = new DTEController(dteConfFileName);
+        }
+        catch (Exception exc) {
+            logger.error("Error initializing DTEController from file: " + dteConfFileName, exc);
+        }
+        gvSvcConfMgr = new ServiceConfigManager();
+        gvSvcConfMgr.setStatisticsDataManager(statisticsDataManager);
+        gvContext = new InvocationContext();
+        gvContext.setGVServiceConfigManager(gvSvcConfMgr);
+        gvContext.setStatisticsDataManager(statisticsDataManager);
+        gvContext.setExtraField("DTE_CONTROLLER", dteController);
+    }
+
+    /**
+     * @param gvBuffer
+     * @throws GVCoreException
+     */
+    private void setInvocationContext(GVBuffer gvBuffer) throws GVCoreException {
+        if (isSingleThread) {
+            return;
+        }
+        if (gvBuffer != null) {
+            gvContext.setContext(flowName, gvBuffer);
+        }
+        gvContext.push();
+    }
+
+    private void resetInvocationContext() {
+        if (!isSingleThread) {
+            try {
+                gvContext.pop();
+                gvContext.cleanup();
+            }
+            catch (Exception exc) {
+                logger.warn("Failed cleanUp() of InvocationContext on GVSubFlow " + this.flowName, exc);
+            }
         }
     }
 
@@ -247,16 +304,18 @@ public class GVSubFlow
      * @throws GVCoreException
      *         if errors occurs
      */
-    public GVBuffer internalPerform(GVBuffer gvBuffer, boolean onDebug) throws GVCoreException
+    public GVBuffer internalPerform(GVBuffer gvBuffer, boolean onDebug) throws GVCoreException, InterruptedException
     {
         Level level = null;
+        boolean success = false;
+        String inID = gvBuffer.getId().toString();
         try {
             level = GVLogger.setThreadMasterLevel(loggerLevel);
 
             if (logger.isDebugEnabled()) {
                 logger.debug(GVFormatLog.formatBEGIN(flowName, gvBuffer));
             }
-            getGVOperationInfo();
+            getGVSubFlowInfo();
     
             GVFlowNode flowNode = flowNodes.get(firstNode);
             if (flowNode == null) {
@@ -268,16 +327,17 @@ public class GVSubFlow
             Map<String, Object> environment = new HashMap<String, Object>();
             environment.put(flowNode.getInput(), gvBuffer);
             String nextNode = firstNode;
-    
+
             if (onDebug) {
-                String inID = gvBuffer.getId().toString();
                 DebugSynchObject synchObj = DebugSynchObject.getSynchObject(inID, null);
                 ExecutionInfo parent = synchObj.getExecutionInfo();
                 ExecutionInfo info = new ExecutionInfo(parent);
                 info.setSubflow(flowName);
                 synchObj.setExecutionInfo(info);
-                while (!nextNode.equals("")) {
-                    operationInfo.setFlowStatus(inID, flowName, nextNode);
+                while (!nextNode.equals("") && !isInterrupted()) {
+                    if (subflowInfo != null) {
+                        subflowInfo.setFlowStatus(inID, nextNode);
+                    }
                     flowNode = flowNodes.get(nextNode);
                     if (flowNode == null) {
                         logger.error("FlowNode " + nextNode + " not configured. Check configuration.");
@@ -290,11 +350,10 @@ public class GVSubFlow
                 synchObj.terminated();
             }
             else {
-                while (!nextNode.equals("")) {
-                    /*
-                     * if (operationInfo != null) {
-                     * operationInfo.setFlowStatus(inID, nextNode); }
-                     */
+                while (!nextNode.equals("") && !isInterrupted()) {
+                    if (subflowInfo != null) {
+                        subflowInfo.setFlowStatus(inID, nextNode);
+                    }
                     flowNode = flowNodes.get(nextNode);
                     if (flowNode == null) {
                         logger.error("FlowNode " + nextNode + " not configured. Check configuration.");
@@ -305,6 +364,10 @@ public class GVSubFlow
                 }
             }
     
+            if (isInterrupted()) {
+                throw new InterruptedException("Subflow [" + flowName + "] interrupted!");
+            }
+
             Object output = environment.get(flowNode.getOutput());
             if (output instanceof Throwable) {
                 if (output instanceof GVCoreException) {
@@ -317,27 +380,35 @@ public class GVSubFlow
             if (logger.isDebugEnabled()) {
                 logger.debug(GVFormatLog.formatEND(flowName, (GVBuffer) output));
             }
-    
+            success = true;
+
             return (GVBuffer) output;
         }
+        catch (InterruptedException exc) {
+            logger.error("Subflow [" + flowName + "] interrupted!", exc);
+            throw exc;
+        }
         finally {
+            if (subflowInfo != null) {
+                subflowInfo.flowTerminated(inID, success);
+            }
             GVLogger.removeThreadMasterLevel(level);
         }
     }
 
     /**
-     * Initialize the associated OperationInfo instance
+     * Initialize the associated SubFlowInfo instance
      */
-    private void getGVOperationInfo()
+    private void getGVSubFlowInfo()
     {
-        if (operationInfo == null) {
+        if (subflowInfo == null) {
             try {
-                operationInfo = ServiceOperationInfoManager.instance().getOperationInfo(serviceName, operationName,
-                        true);
+                subflowInfo = ServiceOperationInfoManager.instance().getSubFlowInfo(serviceName, operationName, 
+                        flowName, true);
             }
             catch (Exception exc) {
                 logger.warn("Error on MBean registration: " + exc);
-                operationInfo = null;
+                subflowInfo = null;
             }
         }
     }
@@ -346,6 +417,10 @@ public class GVSubFlow
      * @return the actual logger level
      */
     public Level getLoggerLevel() {
+        getGVSubFlowInfo();
+        if (subflowInfo != null) {
+            return subflowInfo.getLoggerLevelj();
+        }
         return this.loggerLevel;
     }
 
@@ -377,15 +452,7 @@ public class GVSubFlow
                     logger.warn("Failed cleanUp() operation on GVFlowNode " + node.getId(), exc);
                 }
             }
-            if (!isSingleThread) {
-                try {
-                    gvContext.pop();
-                    gvContext.cleanup();
-                }
-                catch (Exception exc) {
-                    logger.warn("Failed cleanUp() of InvocationContext on GVSubFlow " + this.flowName, exc);
-                }
-            }
+            resetInvocationContext();
         }
         finally {
             GVLogger.removeThreadMasterLevel(level);
@@ -407,12 +474,31 @@ public class GVSubFlow
         }
         flowNodes.clear();
         if (!isSingleThread) {
-            gvContext.destroy();
-            gvContext = null;
+            if (gvContext != null) {
+                gvContext.destroy();
+                gvContext = null;
+            }
+            if (statisticsDataManager != null) {
+                statisticsDataManager.destroy();
+                statisticsDataManager = null;
+            }
+            if (gvSvcConfMgr != null) {
+                gvSvcConfMgr.destroy();
+                gvSvcConfMgr = null;
+            }
             if (dteController != null) {
                 dteController.destroy();
                 dteController = null;
             }
         }
+    }
+    
+    /**
+     * 
+     * @return
+     *        the current Thread interrupted state
+     */
+    protected boolean isInterrupted() {
+        return Thread.currentThread().isInterrupted();
     }
 }
